+++ lib/better_player.dart
@@ -20,6 +20,7 @@ export 'src/list/better_player_list_video_player.dart';
 export 'src/list/better_player_list_video_player_controller.dart';
 export 'src/playlist/better_player_playlist.dart';
 export 'src/playlist/better_player_playlist_configuration.dart';
+export 'src/playlist/better_player_playlist_controller.dart';
 export 'src/subtitles/better_player_subtitles_configuration.dart';
 export 'src/subtitles/better_player_subtitles_source.dart';
 export 'src/subtitles/better_player_subtitles_source_type.dart';